Clear up argument/parameter confusion in step 3
Awesome guide. I'm not sure if the term "parameter" is explicitly being used here for the teaching purposes, but it would probably be more correct to use the term "argument". The formal definition is that *parameters* are the named variables inside a function and *arguments* are the values you call the function with.